Implement permissions (#20)
* Only show export button when there's data to export

* Tests for waitlist deletion, readme updated, installed dbal

* Tests added for waitlist policy, started subscriber policy

* Fix travisci

* Test subscriber deletion

* Fixed bug after deleting subscriber

* Redirect from default laravel welcome page, fix view bug

* Fixed failing test

* Share permissions with frontend

* Higher order function for permissions

+++ app/Http/Controllers/SubscribersController.php
@@ -2,30 +2,27 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Resources\WaitlistResource;
 use App\Models\Subscriber;
-use Illuminate\Support\Facades\Session;
-use Inertia\Inertia;
 
 class SubscribersController extends Controller
 {
     public function delete(Subscriber $subscriber)
     {
-        // todo: permissions
+        if (!auth()->user()->can('delete', $subscriber)) {
+            return response('You are not allowed to delete this subscriber!', 403);
+        }
 
         Subscriber::where('referrer_id', $subscriber->id)
             ->update(['referrer_id' => null]);
 
+        $waitlistId = $subscriber->waitlist->id;
+
         $subscriber->delete();
 
-        Inertia::share('flash', function () {
-            return [
-                'successMessage' => 'Successfully deleted subscriber!',
-            ];
-        });
-
-        return Inertia::render('Lists/Show', [
-            'list' => new WaitlistResource($subscriber->waitlist),
-        ]);
+        return response()
+            ->redirectToRoute('lists.show', [
+                'waitlist' => $waitlistId,
+            ])
+            ->with('successMessage', 'Successfully deleted subscriber!');
     }
 }
